requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
asyncer==0.0.8
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
cleanlab==2.7.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==2.19.0
debugpy==1.8.14
decorator==5.2.1
deepeval==3.3.3
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dspy==2.6.27
et-xmlfile==2.0.0
execnet==2.1.1
executing==2.2.0
fakeredis==2.30.3
fast-fit==1.2.1
fastapi==0.116.1
fastexcel==0.14.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.3.1
gdown==5.2.0
google-auth==2.40.3
google-genai==1.27.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpclib==0.4.7
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.2
hyperframe==6.1.0
identify==2.6.12
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib-metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json-repair==0.48.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
litellm==1.74.8
magicattr==0.1.6
mako==1.3.10
markdown==3.8.2
markdown-it-py==3.0.0
markdownify==1.1.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
modal==1.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nlpaug==1.1.11
nltk==3.9.1
nodeenv==1.9.1
notebook==7.4.3
notebook-shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.5.1
openai==1.97.1
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
optuna==4.4.0
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
polars==1.31.0
portalocker==3.2.0
posthog==3.25.0
pre-commit==4.2.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyfiglet==1.0.3
pygments==2.19.1
pyparsing==3.2.3
pysocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.16.0
seaborn==0.13.2
send2trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.34.1
setuptools==80.9.0
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
statsmodels==0.14.5
sympy==1.13.1
synchronicity==0.10.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
torch==2.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
triton==3.2.0
typer==0.16.0
types-certifi==2021.10.8.3
types-python-dateutil==2.9.0.20250516
types-toml==0.10.8.20240310
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.5.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
